requirements.txt
# DesignApp — Python Geotechnical Analysis Suite
# Python 3.12+
# Install: pip install -r requirements.txt

# Web framework
flask>=3.0

# PDF export
reportlab>=4.0

# Word export
python-docx>=1.1

# Plot / figure export
matplotlib>=3.8
numpy>=1.26

# PDF merge (project report, test content extraction)
pypdf>=4.0

# Testing
pytest>=8.0

# Desktop UI (Phase 6)
PySide6>=6.6
